refactor: Lotto클래스 위치 변경
